# src/main/java/com/witboost/plugin/informatica/datacatalog/controller/DataCatalogPluginController.java
package com.witboost.plugin.informatica.datacatalog.controller;

import com.witboost.plugin.informatica.datacatalog.openapi.api.DataCatalogPluginApiDelegate;
import com.witboost.plugin.informatica.datacatalog.openapi.model.EntityReference;
import com.witboost.plugin.informatica.datacatalog.openapi.model.ProvisioningRequest;
import com.witboost.plugin.informatica.datacatalog.openapi.model.ProvisioningStatus;
import com.witboost.plugin.informatica.datacatalog.openapi.model.ValidationResult;
import com.witboost.plugin.informatica.datacatalog.service.DataCatalogService;
import java.util.Optional;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Service;
import org.springframework.web.context.request.NativeWebRequest;

/**
 * API Controller for the Data Catalog Plugin which implements the autogenerated {@link
 * DataCatalogPluginApiDelegate} interface. The interface defaults the endpoints to throw 501 Not
 * Implemented unless overridden in this class.
 *
 * <p>Exceptions thrown will be handled by {@link DataCatalogPluginExceptionHandler}
 */
@Service("dataCatalogPluginApiDelegate")
public class DataCatalogPluginController implements DataCatalogPluginApiDelegate {

    private final DataCatalogService datacatalogService;

    public DataCatalogPluginController(DataCatalogService datacatalogService) {
        this.datacatalogService = datacatalogService;
    }

    @Deprecated(
            since = "Do not use, overridden here to fix unrelated inherited types error.",
            forRemoval = true)
    @Override
    public Optional<NativeWebRequest> getRequest() {
        return DataCatalogPluginApiDelegate.super.getRequest();
    }

    @Override
    public ResponseEntity<EntityReference> getEntityReference(String componentId) throws Exception {
        return ResponseEntity.ok(this.datacatalogService.getEntityReference(componentId));
    }

    @Override
    public ResponseEntity<ProvisioningStatus> getStatus(String token) throws Exception {
        // TODO implement async
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

    @Override
    public ResponseEntity<ProvisioningStatus> provision(ProvisioningRequest provisioningRequest)
            throws Exception {
        return ResponseEntity.ok(this.datacatalogService.provision(provisioningRequest));
    }

    @Override
    public ResponseEntity<ProvisioningStatus> unprovision(ProvisioningRequest provisioningRequest)
            throws Exception {
        return ResponseEntity.ok(this.datacatalogService.unprovision(provisioningRequest));
    }

    @Override
    public ResponseEntity<ValidationResult> validate(ProvisioningRequest provisioningRequest)
            throws Exception {
        return ResponseEntity.ok(datacatalogService.validate(provisioningRequest));
    }
}