# src/main/java/com/witboost/plugin/informatica/marketplace/controller/MarketplacePluginController.java
package com.witboost.plugin.informatica.marketplace.controller;

import com.witboost.plugin.informatica.marketplace.openapi.api.MarketplaceProxyApiDelegate;
import com.witboost.plugin.informatica.marketplace.openapi.model.ProvisioningRequest;
import com.witboost.plugin.informatica.marketplace.openapi.model.ProvisioningResultRequest;
import com.witboost.plugin.informatica.marketplace.openapi.model.ProvisioningResultResponse;
import com.witboost.plugin.informatica.marketplace.openapi.model.UpdateAclRequest;
import com.witboost.plugin.informatica.marketplace.openapi.model.ValidationResult;
import com.witboost.plugin.informatica.marketplace.service.MarketplaceService;
import java.util.Optional;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Service;
import org.springframework.web.context.request.NativeWebRequest;

/**
 * API Controller for the Marketplace Plugin which implements the autogenerated {@link
 * MarketplaceProxyApiDelegate}. The interface defaults the endpoints to throw 501 Not Implemented
 * unless overridden in this class.
 *
 * <p>Exceptions thrown will be handled by {@link MarketplacePluginExceptionHandler}
 */
@Service("marketplaceProxyApiDelegate")
public class MarketplacePluginController implements MarketplaceProxyApiDelegate {

    private final MarketplaceService marketplaceService;

    public MarketplacePluginController(MarketplaceService marketplaceService) {
        this.marketplaceService = marketplaceService;
    }

    @Override
    public Optional<NativeWebRequest> getRequest() {
        return MarketplaceProxyApiDelegate.super.getRequest();
    }

    @Override
    public ResponseEntity<String> delete(ProvisioningResultRequest provisioningResultRequest)
            throws Exception {
        return ResponseEntity.ok(this.marketplaceService.delete(provisioningResultRequest));
    }

    @Override
    public ResponseEntity<ProvisioningResultResponse> insertProvisioningResults(
            ProvisioningResultRequest provisioningResultRequest) throws Exception {
        return ResponseEntity.ok(
                this.marketplaceService.insertProvisioningResults(provisioningResultRequest));
    }

    @Override
    public ResponseEntity<String> updateAcl(UpdateAclRequest updateAclRequest) throws Exception {
        return ResponseEntity.ok(this.marketplaceService.updateAcl(updateAclRequest));
    }

    @Override
    public ResponseEntity<ValidationResult> validate(ProvisioningRequest provisioningRequest)
            throws Exception {
        return ResponseEntity.ok(this.marketplaceService.validate(provisioningRequest));
    }
}